fixed some bugs & tested others, still some problems with the compId asignment in different odds sites

+++ src/extra/StandartResponses.java
@@ -3,9 +3,9 @@ package extra;
 public class StandartResponses {
 	public static String NOT_FOUND = "NOT_FOUND";
 
-	public static int LEV_DISTANCE = 2;
-	public static int DISMIS_WORD = 2;
-	public static int TEAM_DIST = 3;
+	public static float LEV_DISTANCE = 2;
+	public static float DISMIS_WORD = 2;
+	public static float TEAM_DIST = 3;
 	
 	public static int DEFAULT_NULL =-1;
 	public static int ALL_MATCHES_TAKEN =-2;
